+++ __mocks__/game.ts
@@ -1,24 +1,32 @@
 /**
  * This file mocks the FoundryVTT game global so that it can be used in testing
  */
-import {jest} from '@jest/globals';
+import { jest } from "@jest/globals";
 //@ts-ignore
 const local: Localization = {
-    lang: '',
+    lang: "",
     translations: {},
-    initialize: () => {return Promise.resolve();},
-    localize: (stringId: string) => {return '';},
-    format: (stringId: string, replacements: any) => {return '';},
-    setLanguage: (lang: string) => {return Promise.resolve();}
+    initialize: () => {
+        return Promise.resolve();
+    },
+    localize: (stringId: string) => {
+        return "";
+    },
+    format: (stringId: string, replacements: any) => {
+        return "";
+    },
+    setLanguage: (lang: string) => {
+        return Promise.resolve();
+    }
 };
 
 // @ts-ignore
 const user: User = {
-    name: '',
-    id: '',
+    name: "",
+    id: "",
     active: true,
-    viewedScene: '',
-    avatar: '',
+    viewedScene: "",
+    avatar: "",
     //character: {},
     // @ts-ignore
     permissions: [],
@@ -28,18 +36,30 @@ const user: User = {
     //@ts-ignore
     data: {},
     // @ts-ignore
-    can: jest.fn((permission: string) => {return false;}),
-    hasPermission: jest.fn((permission: string) => {return false;}),
+    can: jest.fn((permission: string) => {
+        return false;
+    }),
+    hasPermission: jest.fn((permission: string) => {
+        return false;
+    }),
     // @ts-ignore
-    hasRole: jest.fn((role: string) => {return true;}),
+    hasRole: jest.fn((role: string) => {
+        return true;
+    }),
     // @ts-ignore
-    isRole: jest.fn((role: string) => {return false;}),
+    isRole: jest.fn((role: string) => {
+        return false;
+    }),
     // @ts-ignore
     setPermission: jest.fn((premission: string, allowed: boolean) => {}),
     // @ts-ignore
-    assignHotbarMacro: jest.fn((macro: Macro | null, slot: string | number, {fromSlot}: { fromSlot: number }): Promise<User> => { return Promise.resolve(user);}),
+    assignHotbarMacro: jest.fn((macro: Macro | null, slot: string | number, { fromSlot }: { fromSlot: number }): Promise<User> => {
+        return Promise.resolve(user);
+    }),
     // @ts-ignore
-    getHotbarMacros: jest.fn((page?: number): Macro[] => {return [];})
+    getHotbarMacros: jest.fn((page?: number): Macro[] => {
+        return [];
+    })
 };
 
 // @ts-ignore
@@ -53,7 +73,9 @@ const game = {
         get: jest.fn(),
         register: jest.fn((moduleName: string, settingName: string, data: any) => {}),
         registerMenu: jest.fn(),
-        set: (moduleName: string, settingName: string, data: any) => {return Promise.resolve(true);}
+        set: (moduleName: string, settingName: string, data: any) => {
+            return Promise.resolve(true);
+        }
     },
     time: {
         worldTime: 10,
@@ -65,11 +87,11 @@ const game = {
     },
     combats: {
         size: 0,
-        find: jest.fn((v:Function)=>{
-            return v.call(undefined, {started: true});
+        find: jest.fn((v: Function) => {
+            return v.call(undefined, { started: true });
         }),
         filter: (v: any) => {
-            return v.call(undefined, {started: true});
+            return v.call(undefined, { started: true });
         }
     },
     modules: {
@@ -82,11 +104,11 @@ const game = {
     },
     users: {
         get: jest.fn(),
-        find: (v: any)=>{
-            return v.call(undefined, {isGM: false, active: true});
+        find: (v: any) => {
+            return v.call(undefined, { isGM: false, active: true });
         },
         forEach: (v: any) => {
-            return v.call(undefined, {id: ''});
+            return v.call(undefined, { id: "" });
         },
         filter: (v: any) => {
             return v.call(undefined, user);
@@ -95,11 +117,13 @@ const game = {
             return [v.call(undefined, user)];
         },
         unshift: (v: any) => {},
-        contents: () => {return[user]}
+        contents: () => {
+            return [user];
+        }
     },
     scenes: null,
     system: {
-        id: '',
+        id: "",
         data: {
             version: "1.2.3"
         }
@@ -109,10 +133,21 @@ const game = {
         get: (id: string, obj: any) => {
             if (id) {
                 return {
-                    getFlag: jest.fn().mockReturnValue({calendarId: 'test', startDate: {}, endDate: {}, allDay: true, repeats: 0, order: 0, categories: [], remindUsers: ['qwe']}),
+                    getFlag: jest.fn().mockReturnValue({
+                        calendarId: "test",
+                        startDate: {},
+                        endDate: {},
+                        allDay: true,
+                        repeats: 0,
+                        order: 0,
+                        categories: [],
+                        remindUsers: ["qwe"]
+                    }),
                     update: jest.fn(),
                     delete: async () => {},
-                    testUserPermission: () => {return true;},
+                    testUserPermission: () => {
+                        return true;
+                    },
                     sheet: {
                         delete: jest.fn(),
                         render: jest.fn(),
@@ -122,26 +157,40 @@ const game = {
             }
             return null;
         },
-        forEach: (v: any) => {return v.call(undefined, {})},
+        forEach: (v: any) => {
+            return v.call(undefined, {});
+        },
         directory: {
             folders: {
-                find: (v: any) => {return v.call(undefined, {
-                    getFlag: jest.fn().mockReturnValueOnce(undefined).mockReturnValue({})
-                });}
+                find: (v: any) => {
+                    return v.call(undefined, {
+                        getFlag: jest.fn().mockReturnValueOnce(undefined).mockReturnValue({})
+                    });
+                }
             }
         }
     },
     macros: {
-        forEach: (v: any) => {return v.call(undefined, {canExecute: true, name: 'asd', id: '123'})},
-        get: () => {return {canExecute: true, execute: jest.fn()}}
+        forEach: (v: any) => {
+            return v.call(undefined, { canExecute: true, name: "asd", id: "123" });
+        },
+        get: () => {
+            return { canExecute: true, execute: jest.fn() };
+        }
     },
     video: {
-        getYouTubePlayer: async () => {return {seekTo: () => {}}},
-        isYouTubeURL: (s: string) => {return s.indexOf('youtube') > -1},
-        getYouTubeEmbedURL: () => {return ''}
+        getYouTubePlayer: async () => {
+            return { seekTo: () => {} };
+        },
+        isYouTubeURL: (s: string) => {
+            return s.indexOf("youtube") > -1;
+        },
+        getYouTubeEmbedURL: () => {
+            return "";
+        }
     },
     world: {
-        id: 'worldId'
+        id: "worldId"
     },
     release: {
         generation: 11
@@ -156,8 +205,8 @@ global.game = game;
 
 // @ts-ignore
 global.ui = {
-    chat:{
-        _lastId: '',
+    chat: {
+        _lastId: "",
         _state: 0,
         render: jest.fn((force: boolean) => {})
     },
@@ -165,8 +214,15 @@ global.ui = {
         info: jest.fn((message: string) => {}),
         warn: jest.fn((message: string) => {}),
         error: jest.fn((message: string) => {})
-    }
+    },
+    windows: {},
+    activeWindow: null
 };
 
 // @ts-ignore
-global.getRoute = (a: string) => {return a;};
+global.getRoute = (a: string) => {
+    return a;
+};
+
+//@ts-ignore
+global._maxZ = 100;