+++ src/classes/chat/chat-timestamp.test.ts
@@ -40,7 +40,7 @@ describe("Chat Timestamp Tests", () => {
 
     test("Add Game Time To Message", () => {
         const cm = {
-            isOwner: true,
+            isAuthor: true,
             setFlag: jest.fn((mn: string, key: string, data: any) => {
                 cm.flags[mn] = {};
                 cm.flags[mn][key] = data;
@@ -66,7 +66,7 @@ describe("Chat Timestamp Tests", () => {
         ChatTimestamp.addGameTimeToMessage(cm);
         expect(cm.setFlag).toHaveBeenCalledTimes(1);
 
-        cm.isOwner = false;
+        cm.isAuthor = false;
         //@ts-ignore
         ChatTimestamp.addGameTimeToMessage(cm);
         expect(cm.setFlag).toHaveBeenCalledTimes(1);